Add files via upload
The size of the files was higher than Github allowed to be uploaded so had to split the rar file into multiple volumes. Extract all in a single folder

In order for Plugins to be found, they must be located in one of the search paths for Plugins, either in your project (game folder), or in the Engine itself.

Search Path For Engine

/[UE4 Root]/Engine/Plugins/[Plugin Name]/

Search Path For Game

/[Project Root]/Plugins/[Plugin Name]/

Either copy the extracted "Plugins" folder into any one of the above mentioned directory or if a "Plugins" folder already exists in the mentioned paths (which it probably will) then copy the folder "AdCollection" located inside the extracted "Plugins" folder to the existing "Plugins" folder inside the above mentioned paths.